+++ sample/config.js
@@ -1,0 +1,168 @@
+System.config({
+  baseURL: "/",
+  defaultJSExtensions: true,
+  transpiler: "babel",
+  babelOptions: {
+    "optional": [
+      "runtime",
+      "optimisation.modules.system",
+      "es7.decorators",
+      "es7.classProperties"
+    ]
+  },
+  paths: {
+    "github:*": "jspm_packages/github/*",
+    "npm:*": "jspm_packages/npm/*"
+  },
+
+  packages: {
+    "/aurelia-validation": {
+      "main": "index",
+      "format": "amd",
+      "dependencies": {
+        "aurelia-binding": "github:aurelia/binding@^0.9.1",
+        "aurelia-dependency-injection": "github:aurelia/dependency-injection@^0.10.1",
+        "aurelia-templating": "github:aurelia/templating@^0.15.3"
+      }
+    }
+  },
+
+  map: {
+    "aurelia-binding": "github:aurelia/binding@0.9.1",
+    "aurelia-bootstrapper": "github:aurelia/bootstrapper@0.17.0",
+    "aurelia-metadata": "github:aurelia/metadata@0.8.0",
+    "aurelia-templating": "github:aurelia/templating@0.15.3",
+    "babel": "npm:babel-core@5.8.23",
+    "babel-runtime": "npm:babel-runtime@5.8.20",
+    "core-js": "npm:core-js@1.1.4",
+    "font-awesome": "npm:font-awesome@4.4.0",
+    "twbs/bootstrap": "github:twbs/bootstrap@3.3.5",
+    "github:aurelia/binding@0.9.1": {
+      "aurelia-dependency-injection": "github:aurelia/dependency-injection@0.10.1",
+      "aurelia-metadata": "github:aurelia/metadata@0.8.0",
+      "aurelia-task-queue": "github:aurelia/task-queue@0.7.0",
+      "core-js": "npm:core-js@0.9.18"
+    },
+    "github:aurelia/bootstrapper@0.17.0": {
+      "aurelia-event-aggregator": "github:aurelia/event-aggregator@0.8.0",
+      "aurelia-framework": "github:aurelia/framework@0.16.0",
+      "aurelia-history": "github:aurelia/history@0.7.0",
+      "aurelia-history-browser": "github:aurelia/history-browser@0.8.0",
+      "aurelia-loader-default": "github:aurelia/loader-default@0.10.0",
+      "aurelia-logging-console": "github:aurelia/logging-console@0.7.1",
+      "aurelia-router": "github:aurelia/router@0.12.0",
+      "aurelia-templating": "github:aurelia/templating@0.15.3",
+      "aurelia-templating-binding": "github:aurelia/templating-binding@0.15.0",
+      "aurelia-templating-resources": "github:aurelia/templating-resources@0.15.2",
+      "aurelia-templating-router": "github:aurelia/templating-router@0.16.1",
+      "core-js": "npm:core-js@0.9.18"
+    },
+    "github:aurelia/dependency-injection@0.10.1": {
+      "aurelia-logging": "github:aurelia/logging@0.7.0",
+      "aurelia-metadata": "github:aurelia/metadata@0.8.0",
+      "core-js": "npm:core-js@0.9.18"
+    },
+    "github:aurelia/event-aggregator@0.8.0": {
+      "aurelia-logging": "github:aurelia/logging@0.7.0"
+    },
+    "github:aurelia/framework@0.16.0": {
+      "aurelia-binding": "github:aurelia/binding@0.9.1",
+      "aurelia-dependency-injection": "github:aurelia/dependency-injection@0.10.1",
+      "aurelia-loader": "github:aurelia/loader@0.9.0",
+      "aurelia-logging": "github:aurelia/logging@0.7.0",
+      "aurelia-metadata": "github:aurelia/metadata@0.8.0",
+      "aurelia-path": "github:aurelia/path@0.9.0",
+      "aurelia-task-queue": "github:aurelia/task-queue@0.7.0",
+      "aurelia-templating": "github:aurelia/templating@0.15.3",
+      "core-js": "npm:core-js@0.9.18"
+    },
+    "github:aurelia/history-browser@0.8.0": {
+      "aurelia-history": "github:aurelia/history@0.7.0",
+      "core-js": "npm:core-js@0.9.18"
+    },
+    "github:aurelia/loader-default@0.10.0": {
+      "aurelia-loader": "github:aurelia/loader@0.9.0",
+      "aurelia-metadata": "github:aurelia/metadata@0.8.0"
+    },
+    "github:aurelia/loader@0.9.0": {
+      "aurelia-html-template-element": "github:aurelia/html-template-element@0.3.0",
+      "aurelia-metadata": "github:aurelia/metadata@0.8.0",
+      "aurelia-path": "github:aurelia/path@0.9.0",
+      "core-js": "github:zloirock/core-js@0.8.4"
+    },
+    "github:aurelia/logging-console@0.7.1": {
+      "aurelia-logging": "github:aurelia/logging@0.7.0"
+    },
+    "github:aurelia/metadata@0.8.0": {
+      "core-js": "npm:core-js@0.9.18"
+    },
+    "github:aurelia/route-recognizer@0.7.0": {
+      "aurelia-path": "github:aurelia/path@0.9.0",
+      "core-js": "npm:core-js@0.9.18"
+    },
+    "github:aurelia/router@0.12.0": {
+      "aurelia-dependency-injection": "github:aurelia/dependency-injection@0.10.1",
+      "aurelia-event-aggregator": "github:aurelia/event-aggregator@0.8.0",
+      "aurelia-history": "github:aurelia/history@0.7.0",
+      "aurelia-logging": "github:aurelia/logging@0.7.0",
+      "aurelia-path": "github:aurelia/path@0.9.0",
+      "aurelia-route-recognizer": "github:aurelia/route-recognizer@0.7.0",
+      "core-js": "npm:core-js@0.9.18"
+    },
+    "github:aurelia/templating-binding@0.15.0": {
+      "aurelia-binding": "github:aurelia/binding@0.9.1",
+      "aurelia-logging": "github:aurelia/logging@0.7.0",
+      "aurelia-templating": "github:aurelia/templating@0.15.3"
+    },
+    "github:aurelia/templating-resources@0.15.2": {
+      "aurelia-binding": "github:aurelia/binding@0.9.1",
+      "aurelia-dependency-injection": "github:aurelia/dependency-injection@0.10.1",
+      "aurelia-loader": "github:aurelia/loader@0.9.0",
+      "aurelia-logging": "github:aurelia/logging@0.7.0",
+      "aurelia-path": "github:aurelia/path@0.9.0",
+      "aurelia-task-queue": "github:aurelia/task-queue@0.7.0",
+      "aurelia-templating": "github:aurelia/templating@0.15.3",
+      "core-js": "npm:core-js@0.9.18"
+    },
+    "github:aurelia/templating-router@0.16.1": {
+      "aurelia-dependency-injection": "github:aurelia/dependency-injection@0.10.1",
+      "aurelia-metadata": "github:aurelia/metadata@0.8.0",
+      "aurelia-path": "github:aurelia/path@0.9.0",
+      "aurelia-router": "github:aurelia/router@0.12.0",
+      "aurelia-templating": "github:aurelia/templating@0.15.3"
+    },
+    "github:aurelia/templating@0.15.3": {
+      "aurelia-binding": "github:aurelia/binding@0.9.1",
+      "aurelia-dependency-injection": "github:aurelia/dependency-injection@0.10.1",
+      "aurelia-html-template-element": "github:aurelia/html-template-element@0.3.0",
+      "aurelia-loader": "github:aurelia/loader@0.9.0",
+      "aurelia-logging": "github:aurelia/logging@0.7.0",
+      "aurelia-metadata": "github:aurelia/metadata@0.8.0",
+      "aurelia-path": "github:aurelia/path@0.9.0",
+      "aurelia-task-queue": "github:aurelia/task-queue@0.7.0",
+      "core-js": "npm:core-js@0.9.18"
+    },
+    "github:jspm/nodelibs-process@0.1.1": {
+      "process": "npm:process@0.10.1"
+    },
+    "github:twbs/bootstrap@3.3.5": {
+      "jquery": "github:components/jquery@2.1.4"
+    },
+    "npm:babel-runtime@5.8.20": {
+      "process": "github:jspm/nodelibs-process@0.1.1"
+    },
+    "npm:core-js@0.9.18": {
+      "fs": "github:jspm/nodelibs-fs@0.1.2",
+      "process": "github:jspm/nodelibs-process@0.1.1",
+      "systemjs-json": "github:systemjs/plugin-json@0.1.0"
+    },
+    "npm:core-js@1.1.4": {
+      "fs": "github:jspm/nodelibs-fs@0.1.2",
+      "process": "github:jspm/nodelibs-process@0.1.1",
+      "systemjs-json": "github:systemjs/plugin-json@0.1.0"
+    },
+    "npm:font-awesome@4.4.0": {
+      "css": "github:systemjs/plugin-css@0.1.16"
+    }
+  }
+});