+++ sample/src/main.js
@@ -1,0 +1,10 @@
+import 'github:twbs/bootstrap@3.3.5';
+
+export function configure(aurelia) {
+  aurelia.use
+    .standardConfiguration()
+    .developmentLogging()
+    .plugin('aurelia-validation');
+
+  aurelia.start().then(a => a.setRoot('src/app'));
+}