+++ src/debouncer.js
@@ -1,0 +1,17 @@
+export class Debouncer {
+  constructor(debounceTimeout) {
+    this.currentFunction = null;
+    this.debounceTimeout = debounceTimeout;
+  }
+  debounce(func) {
+    this.currentFunction = func;
+    setTimeout(() => {
+      if (func !== null && func !== undefined) {
+        if (func === this.currentFunction) {
+          this.currentFunction = null;
+          func();
+        }
+      }
+    }, this.debounceTimeout);
+  }
+}