+++ src/resources/da-DK.js
@@ -1,4 +1,4 @@
-import {Utilities} from '../validation/utilities';
+﻿import {Utilities} from '../utilities';
 
 export let data = {
   settings: {
@@ -6,17 +6,17 @@ export let data = {
   },
   messages: {
     'isRequired': 'er påkrævet',
-    'onValidateCallback' : 'ikke gyldig værdi',
+    'onValidateCallback': 'ikke gyldig værdi',
     'AlphaNumericOrWhitespaceValidationRule': (newValue, threshold) => {
       return `kan kun indeholde alfanumeriske tegn eller mellemrum`;
     },
     'AlphaNumericValidationRule': (newValue, threshold) => {
       return `kan kun indeholde alfanumeriske tegn`;
     },
-    'AlphaValidationRule' : (newValue, threshold) => {
+    'AlphaValidationRule': (newValue, threshold) => {
       return `kan kun indeholde bogstaver`;
     },
-    'AlphaOrWhitespaceValidationRule' : (newValue, threshold) => {
+    'AlphaOrWhitespaceValidationRule': (newValue, threshold) => {
       return `kan kun indeholde bogstaver og mellemrum`;
     },
     'BetweenLengthValidationRule': (newValue, threshold) => {
@@ -26,7 +26,7 @@ export let data = {
       return `skal være imellem ${Utilities.getValue(threshold.minimumValue)} og ${Utilities.getValue(threshold.maximumValue)}`;
     },
     'CustomFunctionValidationRule': (newValue, threshold) => {
-      return `ikke en gyldig værdi`
+      return `ikke en gyldig værdi`;
     },
     'DigitValidationRule': (newValue, threshold) => {
       return `kan kun indeholde tal`;
@@ -37,19 +37,19 @@ export let data = {
     'EqualityValidationRule': (newValue, threshold) => {
       return `burde være ${Utilities.getValue(threshold.otherValue)}`;
     },
-    'InEqualityValidationRule' : (newValue, threshold) => {
-        return `kan ikke være ${Utilities.getValue(threshold.otherValue)}`;
+    'InEqualityValidationRule': (newValue, threshold) => {
+      return `kan ikke være ${Utilities.getValue(threshold.otherValue)}`;
     },
     'EqualityWithOtherLabelValidationRule': (newValue, threshold) => {
       return `er ikke ens ${Utilities.getValue(threshold.otherValueLabel)}`;
     },
-    'InEqualityWithOtherLabelValidationRule' : (newValue, threshold) => {
+    'InEqualityWithOtherLabelValidationRule': (newValue, threshold) => {
       return `kan ikke være ens ${Utilities.getValue(threshold.otherValueLabel)}`;
     },
     'InCollectionValidationRule': (newValue, threshold) => {
       return `ikke en gyldig værdi`;
     },
-    'MinimumInclusiveValueValidationRule' : (newValue, threshold) => {
+    'MinimumInclusiveValueValidationRule': (newValue, threshold) => {
       return `skal være ${Utilities.getValue(threshold)} eller højere`;
     },
     'MinimumLengthValidationRule': (newValue, threshold) => {
@@ -58,7 +58,7 @@ export let data = {
     'MinimumValueValidationRule': (newValue, threshold) => {
       return `skal være større end ${Utilities.getValue(threshold)}`;
     },
-    'MaximumInclusiveValueValidationRule' : (newValue, threshold) => {
+    'MaximumInclusiveValueValidationRule': (newValue, threshold) => {
       return `skal være ${Utilities.getValue(threshold)} eller mindre`;
     },
     'MaximumLengthValidationRule': (newValue, threshold) => {
@@ -70,7 +70,7 @@ export let data = {
     'NumericValidationRule': (newValue, threshold) => {
       return `skal være en talværdi`;
     },
-    'NoSpacesValidationRule' : (newValue, threshold) => {
+    'NoSpacesValidationRule': (newValue, threshold) => {
       return `må ikke indeholde mellemrum`;
     },
     'RegexValidationRule': (newValue, threshold) => {
@@ -82,11 +82,11 @@ export let data = {
     'StrongPasswordValidationRule': (newValue, threshold) => {
       return `skal være en kombination af små bogstaver, store bogstaver, tal og specialtegn`;
     },
-    'MediumPasswordValidationRule' : (newValue, threshold) => {
+    'MediumPasswordValidationRule': (newValue, threshold) => {
       return `skal indeholde mindst ${Utilities.getValue(threshold)} af de følgende grupper: små bogstaver, store bogstaver, tal og specialtegn`;
     },
-    'URLValidationRule' : (newValue, threshold) =>{
-      return `ikke en gyldig URL`
+    'URLValidationRule': (newValue, threshold) => {
+      return `ikke en gyldig URL`;
     }
   }
 };