+++ test/validationRules/alpha.spec.js
@@ -1,4 +1,4 @@
-import {AlphaValidationRule} from '../../src/validation/validation-rules';
+import {AlphaValidationRule} from '../../src/validation-rules';
 import {Expectations} from '../expectations';
 
 //No need to test empty values, they are filtered out by the "ValidationProperty" depending if they are 'isNotEmpty()'