+++ test/validationRules/email.spec.js
@@ -1,4 +1,4 @@
-import {EmailValidationRule} from '../../src/validation/validation-rules';
+import {EmailValidationRule} from '../../src/validation-rules';
 import {Expectations} from '../expectations';
 
 //No need to test empty values, they are filtered out by the "ValidationProperty" depending if they are 'isNotEmpty()'